requirements.txt
black==21.8b0
click==8.0.1
importlib-metadata==4.8.1
mypy-extensions==0.4.3
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.3.0
regex==2021.8.28
RPi.GPIO==0.7.0
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
zipp==3.5.0
